refactor: amend slr location in satelles module
refactor: amend slr location in satelles module
